// spec/generated/openapi-types.d.ts
export interface components {
    schemas: {
// ...
        AdminShiftSubmissionUpdateRequest: {
            note?: string;
            details?: components["schemas"]["ShiftSubmissionDetail"][];
        };
        EmployeeCreateRequest: {
            employeeCode: string;
            displayName: string;
            department?: string;
            /** Format: date */
            joinedOn?: string;
        };
        EmployeeUpdateRequest: {
            displayName?: string;
            department?: string;
            isActive?: boolean;
            /** Format: date */
            leftOn?: string;
        };
// ...
    };
// ...
}